+++ test/tests/get-schema-for-endpoint.test.ts
@@ -113,7 +113,8 @@ describe('getSchemaForEndpoint', () => {
                 example: '4203312194',
               },
               active_stake: {
-                type: ['string', 'null'],
+                nullable: true,
+                type: 'string',
                 description:
                   'Sum of all the active stakes within the epoch in Lovelaces',
                 example: '784953934049314',
@@ -189,7 +190,8 @@ describe('getSchemaForEndpoint', () => {
               ],
             },
             stake_address: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 'stake1ux3g2c9dx2nhhehyrezyxpkstartcqmu9hk63qgfkccw5rqttygt7',
               description: 'Stake address that controls the key',
@@ -236,13 +238,15 @@ describe('getSchemaForEndpoint', () => {
                 description: {
                   description: 'Description of the stake pool',
                   example: 'The best pool ever',
-                  type: ['string', 'null'],
+                  nullable: true,
+                  type: 'string',
                 },
                 hash: {
                   description: 'Hash of the metadata file',
                   example:
                     '47c0c68cb57f4a5b4a87bad896fc274678e7aea98e200fa14a1cb40c0cab1d8c',
-                  type: ['string', 'null'],
+                  nullable: true,
+                  type: 'string',
                 },
                 hex: {
                   description: 'Hexadecimal pool ID',
@@ -253,12 +257,14 @@ describe('getSchemaForEndpoint', () => {
                 homepage: {
                   description: 'Home page of the stake pool',
                   example: 'https://stakentus.com/',
-                  type: ['string', 'null'],
+                  nullable: true,
+                  type: 'string',
                 },
                 name: {
                   description: 'Name of the stake pool',
                   example: 'Stake Nuts',
-                  type: ['string', 'null'],
+                  nullable: true,
+                  type: 'string',
                 },
                 pool_id: {
                   description: 'Bech32 pool ID',
@@ -269,12 +275,14 @@ describe('getSchemaForEndpoint', () => {
                 ticker: {
                   description: 'Ticker of the stake pool',
                   example: 'NUTS',
-                  type: ['string', 'null'],
+                  nullable: true,
+                  type: 'string',
                 },
                 url: {
                   description: 'URL to the stake pool metadata',
                   example: 'https://stakenuts.com/mainnet.json',
-                  type: ['string', 'null'],
+                  nullable: true,
+                  type: 'string',
                 },
               },
               required: [
@@ -386,7 +394,8 @@ describe('getSchemaForEndpoint', () => {
                 description:
                   'Sum of all the active stakes within the epoch in Lovelaces',
                 example: '784953934049314',
-                type: ['string', 'null'],
+                nullable: true,
+                type: 'string',
               },
             },
             required: [
@@ -425,7 +434,8 @@ describe('getSchemaForEndpoint', () => {
               description: 'Block creation time in UNIX time',
             },
             height: {
-              type: ['integer', 'null'],
+              type: 'integer',
+              nullable: true,
               example: 15243593,
               description: 'Block number',
             },
@@ -436,17 +446,20 @@ describe('getSchemaForEndpoint', () => {
               description: 'Hash of the block',
             },
             slot: {
-              type: ['integer', 'null'],
+              type: 'integer',
+              nullable: true,
               example: 412162133,
               description: 'Slot number',
             },
             epoch: {
-              type: ['integer', 'null'],
+              type: 'integer',
+              nullable: true,
               example: 425,
               description: 'Epoch number',
             },
             epoch_slot: {
-              type: ['integer', 'null'],
+              type: 'integer',
+              nullable: true,
               example: 12,
               description: 'Slot within the epoch',
             },
@@ -468,18 +481,20 @@ describe('getSchemaForEndpoint', () => {
               description: 'Number of transactions in the block',
             },
             output: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example: '128314491794',
               description: 'Total output within the block in Lovelaces',
             },
             fees: {
-              type: ['string', 'null'],
-
+              type: 'string',
+              nullable: true,
               example: '592661',
               description: 'Total fees within the block in Lovelaces',
             },
             block_vrf: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 'vrf_vk1wf2k6lhujezqcfe00l6zetxpnmh9n6mwhpmhm0dvfh3fxgmdnrfqkms8ty',
               description: 'VRF key of the block',
@@ -487,26 +502,30 @@ describe('getSchemaForEndpoint', () => {
               maxLength: 65,
             },
             op_cert: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 'da905277534faf75dae41732650568af545134ee08a3c0392dbefc8096ae177c',
               description:
                 'The hash of the operational certificate of the block producer',
             },
             op_cert_counter: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example: '18',
               description:
                 'The value of the counter used to produce the operational certificate',
             },
             previous_block: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 '43ebccb3ac72c7cebd0d9b755a4b08412c9f5dcb81b8a0ad1e3c197d29d47b05',
               description: 'Hash of the previous block',
             },
             next_block: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 '8367f026cf4b03e116ff8ee5daf149b55ba5a6ec6dec04803b8dc317721d15fa',
               description: 'Hash of the next block',
@@ -623,7 +642,8 @@ describe('getSchemaForEndpoint', () => {
               description: 'Block creation time in UNIX time',
             },
             height: {
-              type: ['integer', 'null'],
+              type: 'integer',
+              nullable: true,
               example: 15243593,
               description: 'Block number',
             },
@@ -634,17 +654,20 @@ describe('getSchemaForEndpoint', () => {
               description: 'Hash of the block',
             },
             slot: {
-              type: ['integer', 'null'],
+              type: 'integer',
+              nullable: true,
               example: 412162133,
               description: 'Slot number',
             },
             epoch: {
-              type: ['integer', 'null'],
+              type: 'integer',
+              nullable: true,
               example: 425,
               description: 'Epoch number',
             },
             epoch_slot: {
-              type: ['integer', 'null'],
+              type: 'integer',
+              nullable: true,
               example: 12,
               description: 'Slot within the epoch',
             },
@@ -666,17 +689,20 @@ describe('getSchemaForEndpoint', () => {
               description: 'Number of transactions in the block',
             },
             output: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example: '128314491794',
               description: 'Total output within the block in Lovelaces',
             },
             fees: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example: '592661',
               description: 'Total fees within the block in Lovelaces',
             },
             block_vrf: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 'vrf_vk1wf2k6lhujezqcfe00l6zetxpnmh9n6mwhpmhm0dvfh3fxgmdnrfqkms8ty',
               description: 'VRF key of the block',
@@ -684,26 +710,30 @@ describe('getSchemaForEndpoint', () => {
               maxLength: 65,
             },
             op_cert: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 'da905277534faf75dae41732650568af545134ee08a3c0392dbefc8096ae177c',
               description:
                 'The hash of the operational certificate of the block producer',
             },
             op_cert_counter: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example: '18',
               description:
                 'The value of the counter used to produce the operational certificate',
             },
             previous_block: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 '43ebccb3ac72c7cebd0d9b755a4b08412c9f5dcb81b8a0ad1e3c197d29d47b05',
               description: 'Hash of the previous block',
             },
             next_block: {
-              type: ['string', 'null'],
+              type: 'string',
+              nullable: true,
               example:
                 '8367f026cf4b03e116ff8ee5daf149b55ba5a6ec6dec04803b8dc317721d15fa',
               description: 'Hash of the next block',